+++ app/App.vue
@@ -1,0 +1,25 @@
+<template>
+  <HelloWorld />
+</template>
+
+<script setup lang="ts">
+import HelloWorld from '@/components/HelloWorld.vue'
+</script>
+
+<style lang="scss">
+body {
+  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen,
+    Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
+  color: var(--base);
+  margin: 0;
+  background-color: #fff;
+  --base: #444;
+  --font-code: 'Source Code Pro', monospace;
+  --color-branding: #3ca877;
+  --color-branding-dark: #416f9c;
+}
+
+#app {
+  height: 100%;
+}
+</style>