+++ app/shims-vue.d.ts
@@ -6,3 +6,14 @@ declare module '*.vue' {
 }
 
 declare module '*.html'
+
+declare module '*.types' {
+  const content: string;
+  export default content;
+}
+
+declare module '*.d.ts' {
+  const content: string;
+  export default content;
+}
+