+++ examples/vite2-demo/src/components/HelloWorld.vue
@@ -66,7 +66,7 @@ export default defineComponent({
 .hello {
   margin: 60px;
   display: grid;
-  grid-template-columns: repeat(2, minmax(0, 1fr));
+  grid-template-columns: repeat(1, minmax(0, 1fr));
   gap: 32px;
 }
 h1 {