+++ examples/vite2-demo/src/main.ts
@@ -1,0 +1,8 @@
+import { createApp } from 'vue'
+import App from './App.vue'
+
+import VueSfcSandbox from 'vue-sfc-sandbox'
+
+createApp(App)
+  .use(VueSfcSandbox)
+  .mount('#app')