+++ examples/vite2-demo/src/shims-vue.d.ts
@@ -1,8 +1,5 @@
-/* eslint-disable */
 declare module '*.vue' {
-  import type { DefineComponent } from 'vue'
+  import { DefineComponent } from 'vue'
   const component: DefineComponent<{}, {}, any>
   export default component
 }
-
-declare module '*.html'