+++ packages/sandbox/utils.ts
@@ -7,3 +7,18 @@ export const debounce = (fn: Function, n = 166) => {
     }, n)
   }
 }
+
+export const logger = (msg: string) => {
+  const styles = [
+    'color: white',
+    'background: #42b983',
+    'margin-left: 4px',
+    'padding: 2px 4px',
+    'border-radius: 2px'
+  ].join(';')
+  console.log(
+    msg,
+    styles,
+    ''
+  )
+}