+++ rollup.config.js
@@ -1,4 +1,4 @@
 // Set options as a parameter, environment variable, or rc file.
-require = require('esm')(module /*, options*/ )
+require = require('esm')(module /*, options */)
 require('esbuild-register')
 module.exports = require('./scripts/rollup.config.ts')