+++ src/constants/default.demo.js
@@ -1,8 +1,0 @@
-export const SFC_CODE_ESM = `<template>
-  <h1>{{ msg }}</h1>
-</template>
-
-<script setup>
-const msg = 'Hello World!'
-</script>
-`